move index to home folder, create state and context files and barrell folder
